feat(logging): split logging_foundation into modules and add structured error logging
Refs #447, #449

- Split lib.rs into init.rs / filter.rs / path.rs / structured.rs
- Add ERROR_KIND_* constants in structured.rs for error_kind field convention
- Replace JA string log in toolpath.rs with key+EN structured fields
- Add ToolPathBuildError local enum to preserve ValidationError type info
- Add error_kind field to cam_sim_visualization_converter warn
- Add logging_foundation dependency to viewmodel/converter
